refactor: move mock splits function to to test file
