+++ internal/split/split_service_test.go
@@ -1,9 +1,12 @@
 package split
 
 import (
+	"io/ioutil"
+	"os"
 	"reflect"
 	"testing"
 
+	log "github.com/sirupsen/logrus"
 	"github.com/splitio/go-client/v6/splitio/client"
 	"github.com/stretchr/testify/require"
 )
@@ -41,6 +44,25 @@ func TestGetAll(t *testing.T) {
 }
 
 func setup() *Service {
+	// Set a custom split config
+	splitConfig.SplitFile = createMockSplits()
 	service, _ := NewService("localhost")
 	return service
+}
+
+// Creates a temporary file with splits used for unit-testing
+func createMockSplits() string {
+	dir, err := os.UserHomeDir()
+	if err != nil {
+		log.Errorf("could not get user home directory: %s", err)
+	}
+	// Create a temporary file that holds test splits for testing purposes
+	blob := []byte(MockSplits)
+	filename := dir + "/mock.split"
+	err = ioutil.WriteFile(filename, blob, 0777)
+	if err != nil {
+		log.Errorf("could not create temp file: %s", err)
+	}
+
+	return filename
 }
